# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Lenny Growth Assistant API
  version: "1.0.0"
  description: >
    Source of truth for api/ request and response contracts (architecture.md §5).
    Pydantic models are derived from this file; a Schemathesis CI gate checks
    the running service against it (§12.6). /internal/retrieve is documented
    here for completeness but is not proxied to the browser.
servers:
  - url: http://localhost:8000

components:
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, trace_id, retryable]
          properties:
            code: { type: string, example: OLLAMA_UNREACHABLE }
            message: { type: string }
            trace_id: { type: string }
            retryable: { type: boolean }

    Session:
      type: object
      required: [id, title, provider, model, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        title: { type: [string, "null"] }
        provider: { type: string, example: ollama }
        model: { type: string, example: "qwen2.5:7b-instruct" }
        enabled_skills:
          type: [array, "null"]
          items: { type: string }
          description: >
            Root CLAUDE.md invariant #4: null = every discovered skill
            active for this session (default); a non-null array is an
            explicit allowlist by skill name. Skills carry no tools, so
            this can only narrow prompt content, never grant capability.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    UpdateSessionCapabilitiesRequest:
      type: object
      properties:
        enabled_skills:
          type: [array, "null"]
          items: { type: string }

    Citation:
      type: object
      required: [chunk_id, episode, guest, youtube_url, rank, score]
      properties:
        chunk_id: { type: integer }
        episode: { type: string }
        guest: { type: string }
        youtube_url: { type: [string, "null"] }
        start_seconds: { type: [integer, "null"], description: "Nearest preceding speaker-turn timestamp; append as &t={start_seconds}s to youtube_url to deep-link." }
        rank: { type: integer, minimum: 1 }
        score: { type: number }

    Message:
      type: object
      required: [id, session_id, role, content, created_at, abstained]
      properties:
        id: { type: string, format: uuid }
        session_id: { type: string, format: uuid }
        role: { type: string, enum: [user, assistant, system] }
        content: { type: string }
        rewritten_query: { type: [string, "null"] }
        trace_id: { type: string }
        provider: { type: [string, "null"] }
        model: { type: [string, "null"] }
        latency_ms: { type: [integer, "null"] }
        abstained: { type: boolean }
        created_at: { type: string, format: date-time }
        citations:
          type: array
          items: { $ref: "#/components/schemas/Citation" }

    SessionDetail:
      allOf:
        - $ref: "#/components/schemas/Session"
        - type: object
          required: [messages]
          properties:
            messages:
              type: array
              items: { $ref: "#/components/schemas/Message" }

    CreateSessionRequest:
      type: object
      properties:
        title: { type: [string, "null"] }

    PostMessageRequest:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1, maxLength: 8000 }

    ChunkDetail:
      type: object
      required: [id, text, episode]
      properties:
        id: { type: integer }
        text: { type: string }
        ordinal: { type: integer }
        episode:
          type: object
          required: [id, guest, title, youtube_url, source_path]
          properties:
            id: { type: integer }
            guest: { type: string }
            title: { type: string }
            youtube_url: { type: [string, "null"] }
            publish_date: { type: [string, "null"], format: date }
            source_path: { type: string }
        start_seconds: { type: [integer, "null"] }

    Artifact:
      type: object
      required: [id, session_id, kind, content, sanitized, created_at]
      properties:
        id: { type: string, format: uuid }
        session_id: { type: string, format: uuid }
        message_id: { type: [string, "null"], format: uuid }
        kind: { type: string, enum: [markdown, html] }
        title: { type: [string, "null"] }
        content: { type: string }
        sanitized: { type: boolean }
        created_at: { type: string, format: date-time }

    ConfigResponse:
      type: object
      required: [provider, model, cloud_available, corpus, capabilities]
      properties:
        provider: { type: string, enum: [ollama, anthropic] }
        model: { type: string }
        cloud_available: { type: boolean }
        corpus:
          type: object
          required: [episode_count, chunk_count]
          properties:
            episode_count: { type: integer }
            chunk_count: { type: integer }
        capabilities: { $ref: "#/components/schemas/Capabilities" }

    Capabilities:
      description: >
        Root CLAUDE.md invariant #4 / architecture.md §8.5: what's actually
        active on the agent right now. `extensions_enabled: false` (the
        default) means Pi's .pi/extensions/ mechanism is off entirely, not
        just empty. `agent_reachable: false` means this snapshot could not
        be fetched from the agent service and the other fields are empty
        defaults, not a confirmed empty state.
      type: object
      required: [skills, extensions, extensions_enabled, tools, agent_reachable]
      properties:
        skills:
          type: array
          items:
            type: object
            required: [name, description]
            properties:
              name: { type: string }
              description: { type: string }
        extensions:
          type: array
          items:
            type: object
            required: [path, tools]
            properties:
              path: { type: string }
              tools: { type: array, items: { type: string } }
        extensions_enabled: { type: boolean }
        tools: { type: array, items: { type: string } }
        agent_reachable: { type: boolean }

    CreateExtensionProposalRequest:
      description: >
        A user-drafted proposal for a new agent/.pi/extensions/ entry.
        Recording one never deploys it (root CLAUDE.md invariant #4) — a
        maintainer must still commit the code, add a matching
        agent/.pi/extensions/manifest.json entry, and pass
        tools/check_extension_manifest.py before it can run.
      type: object
      required: [title, description, tool_names, code]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, minLength: 1, maxLength: 2000 }
        tool_names:
          type: array
          minItems: 1
          maxItems: 20
          items: { type: string, pattern: "^[a-z][a-z0-9_]{0,63}$" }
        code: { type: string, minLength: 1, maxLength: 100000 }

    ExtensionProposal:
      type: object
      required: [id, title, description, tool_names, code, sha256, status, session_id, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        tool_names: { type: array, items: { type: string } }
        code: { type: string }
        sha256:
          type: string
          description: sha256 of `code` — paste straight into manifest.json's matching entry.
        status: { type: string, enum: [pending, approved, rejected] }
        session_id: { type: [string, "null"], format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    ExtensionProposalList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ExtensionProposal" }

    UpdateExtensionProposalStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending, approved, rejected] }

    HealthDeps:
      type: object
      required: [db, ollama, agent]
      properties:
        db: { type: string, enum: [ok, degraded, down] }
        ollama: { type: string, enum: [ok, degraded, down] }
        agent: { type: string, enum: [ok, degraded, down] }

    RetrieveRequest:
      type: object
      required: [query, session_id]
      properties:
        query: { type: string }
        session_id: { type: string, format: uuid }
        k: { type: integer, default: 8 }

    RetrieveResponse:
      type: object
      required: [abstained, chunks]
      properties:
        abstained: { type: boolean }
        floor: { type: number }
        chunks:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/Citation"
              - type: object
                required: [text]
                properties:
                  text: { type: string }

    # Contract addition (agent/ build, architecture.md §8.3): the agent's
    # create_artifact tool needs a write path to persist an artifact before
    # it can emit an `artifact` SSE frame carrying a real artifact_id. The
    # original contract only exposed GET /artifacts/{id}. This is the
    # narrowest addition that closes that gap — see POST /internal/artifacts
    # below. Not exposed to the browser, same trust boundary as
    # /internal/retrieve (Compose-internal network + shared-secret header).
    CreateArtifactRequest:
      type: object
      required: [session_id, kind, content]
      properties:
        session_id: { type: string, format: uuid }
        message_id: { type: [string, "null"], format: uuid }
        kind: { type: string, enum: [markdown, html] }
        title: { type: [string, "null"] }
        content: { type: string, maxLength: 120000 }

    # Contract addition (agent/ build): the agent's new edit_artifact tool
    # revises an artifact it already created, instead of pasting the
    # revised content into the chat reply because it has no way to update
    # what it made. `kind` is deliberately absent — an edit changes
    # content/title on the existing row, never its kind. Same trust
    # boundary and session-scoping as CreateArtifactRequest: api verifies
    # session_id matches the artifact's own session_id before applying the
    # update (see PATCH /internal/artifacts/{artifact_id}).
    UpdateArtifactRequest:
      type: object
      required: [session_id, content]
      properties:
        session_id: { type: string, format: uuid }
        title: { type: [string, "null"] }
        content: { type: string, maxLength: 120000 }

  securitySchemes:
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token

paths:
  /health:
    get:
      summary: Liveness
      operationId: healthLive
      responses:
        "200":
          description: Process is up
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties: { status: { const: ok } }

  /health/deps:
    get:
      summary: Dependency status
      operationId: healthDeps
      responses:
        "200":
          description: Independent status per dependency
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthDeps" }

  /config:
    get:
      summary: Active provider, model, and corpus stats
      operationId: getConfig
      responses:
        "200":
          description: Config snapshot driving the UI provider badge
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigResponse" }

  /sessions:
    post:
      summary: Create a session
      operationId: createSession
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSessionRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
    get:
      summary: List sessions, newest first
      operationId: listSessions
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Paginated session list
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Session" }
                  total: { type: integer }

  /sessions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Session with full message history and citations
      operationId: getSession
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionDetail" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    delete:
      summary: Delete a session and its cascade
      operationId: deleteSession
      responses:
        "204": { description: Deleted }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  /sessions/{id}/capabilities:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      summary: Set this session's active skill allowlist (root CLAUDE.md invariant #4)
      operationId: updateSessionCapabilities
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSessionCapabilitiesRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  /sessions/{id}/messages:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Send a turn. Streams an SSE response (see contracts/sse-frames.schema.json).
      operationId: postMessage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostMessageRequest" }
      responses:
        "200":
          description: >
            text/event-stream of typed frames: stage, token, citation,
            artifact, error, done.
          content:
            text/event-stream:
              schema: { type: string }
        "503":
          description: Provider unavailable (e.g. Ollama unreachable) — no silent failover
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  /extension-proposals:
    get:
      summary: List extension proposals, newest first
      operationId: listExtensionProposals
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExtensionProposalList" }
    post:
      summary: Draft a new extension proposal (review queue only — never auto-deployed)
      operationId: createExtensionProposal
      parameters:
        - name: session_id
          in: query
          required: false
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateExtensionProposalRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExtensionProposal" }

  /extension-proposals/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      summary: Update review status (bookkeeping only — see ExtensionProposal description)
      operationId: updateExtensionProposalStatus
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateExtensionProposalStatusRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExtensionProposal" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  /chunks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      summary: Verbatim chunk text plus episode metadata (serves the citation-expand flow, F2)
      operationId: getChunk
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChunkDetail" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  /artifacts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Fetch a stored artifact
      operationId: getArtifact
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Artifact" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  /internal/retrieve:
    post:
      summary: >
        Vector search with session boost and relevance floor. Caller: agent
        service only. Not proxied to the browser; Compose-internal network
        plus a shared-secret header.
      operationId: internalRetrieve
      security:
        - internalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RetrieveRequest" }
      responses:
        "200":
          description: Ranked chunks with scores and episode metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RetrieveResponse" }
        "401":
          description: Missing or invalid internal token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  /internal/artifacts/{artifact_id}:
    patch:
      summary: >
        Revise an artifact the agent's edit_artifact tool already created
        earlier in the session, in place — no version history, matching the
        no-undo semantics artifacts already have on session delete. Caller:
        agent service only. `session_id` in the body must match the
        artifact's own session_id or the request is rejected as not found,
        the same information-minimal failure shape used elsewhere on this
        internal surface.
      operationId: internalUpdateArtifact
      security:
        - internalToken: []
      parameters:
        - name: artifact_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateArtifactRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Artifact" }
        "401":
          description: Missing or invalid internal token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "404":
          description: No artifact with this id in this session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  /internal/artifacts:
    post:
      summary: >
        Persist an artifact produced by the agent's create_artifact tool
        (contract addition — see the note on CreateArtifactRequest above).
        Caller: agent service only. api is responsible for sanitisation
        (architecture.md §7.6) before the content is ever served to the
        browser; `sanitized` on the response reflects whether that already
        ran synchronously.
      operationId: internalCreateArtifact
      security:
        - internalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateArtifactRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Artifact" }
        "401":
          description: Missing or invalid internal token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }